# swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "This is a Football Statistics API",
    "version" : "1.0.0",
    "title" : "Football Statistics API",
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host" : "calm-forest-73385.herokuapp.com",
  "basePath" : "/",
  "tags" : [ {
    "name" : "developers",
    "description" : "Operations available to regular developers"
  } ],
  "schemes" : [ "https" ],
  "paths" : {
    "/competitions" : {
      "get" : {
        "tags" : [ "developers" ],
        "summary" : "get all competitions",
        "description" : "retrieves all the available competitions.",
        "operationId" : "getCompetitions",
        "produces" : [ "application/json" ],
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "retrieve all available competitions",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/CompetitionItem"
              }
            }
          }
        }
      }
    },
    "/ranking" : {
      "get" : {
        "tags" : [ "developers" ],
        "summary" : "get season's ranking",
        "description" : "By passing the right parameters, you get the ranking of the French football championship (called \"Ligue 1\"\") for the season of your choice.\n",
        "operationId" : "getRanking",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "season",
          "in" : "query",
          "description" : "pass a season to get the ranking for this season",
          "required" : true,
          "type" : "integer",
          "maximum" : 2017,
          "minimum" : 1933,
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "ranking matching criteria",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/RankingItem"
              }
            }
          },
          "400" : {
            "description" : "bad input parameter"
          }
        }
      }
    },
    "/history" : {
      "get" : {
        "tags" : [ "developers" ],
        "summary" : "get opposition history",
        "description" : "get opposition history",
        "operationId" : "getHistory",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "teamHome",
          "in" : "query",
          "description" : "pass the home team",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "teamAway",
          "in" : "query",
          "description" : "pass the visitor team",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "ranking matching criteria",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/HistoryItem"
              }
            }
          },
          "400" : {
            "description" : "bad input parameter"
          }
        }
      }
    }
  },
  "definitions" : {
    "RankingItem" : {
      "type" : "object",
      "required" : [ "draw", "goalDifference", "goals", "goalsAgainst", "goalsAway", "goalsAwayAgainst", "goalsHome", "goalsHomeAgainst", "lost", "matchs", "points", "teamName", "win" ],
      "properties" : {
        "teamName" : {
          "type" : "string",
          "example" : "Olympique Lyonnais"
        },
        "points" : {
          "type" : "integer",
          "format" : "int32",
          "example" : 23
        },
        "matchs" : {
          "type" : "integer",
          "format" : "int32",
          "example" : 38
        },
        "win" : {
          "type" : "integer",
          "format" : "int32",
          "example" : 5
        },
        "lost" : {
          "type" : "integer",
          "format" : "int32",
          "example" : 30
        },
        "draw" : {
          "type" : "integer",
          "format" : "int32",
          "example" : 3
        },
        "goals" : {
          "type" : "integer",
          "format" : "int32",
          "example" : 40
        },
        "goalsAgainst" : {
          "type" : "integer",
          "format" : "int32",
          "example" : 50
        },
        "goalsHome" : {
          "type" : "integer",
          "format" : "int32",
          "example" : 30
        },
        "goalsHomeAgainst" : {
          "type" : "integer",
          "format" : "int32",
          "example" : 10
        },
        "goalsAway" : {
          "type" : "integer",
          "format" : "int32",
          "example" : 10
        },
        "goalsAwayAgainst" : {
          "type" : "integer",
          "format" : "int32",
          "example" : 40
        },
        "goalDifference" : {
          "type" : "integer",
          "format" : "int32",
          "example" : -10
        }
      },
      "example" : {
        "teamName" : "Olympique Lyonnais",
        "matchs" : 38,
        "goalsAway" : 10,
        "draw" : 3,
        "points" : 23,
        "goalsAwayAgainst" : 40,
        "lost" : 30,
        "goalsHome" : 30,
        "goalsHomeAgainst" : 10,
        "goalsAgainst" : 50,
        "win" : 5,
        "goalDifference" : -10,
        "goals" : 40
      }
    },
    "HistoryItem" : {
      "type" : "object",
      "required" : [ "FT", "date", "division", "goaldif", "hgoal", "home", "id", "result", "season", "tier", "totgoal", "vgoal", "visitor" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32",
          "example" : 1
        },
        "date" : {
          "type" : "string",
          "example" : "1956-12-25T00:00:00.000Z"
        },
        "season" : {
          "type" : "integer",
          "format" : "int32",
          "example" : 1956
        },
        "home" : {
          "type" : "string",
          "example" : "Stade Rennes"
        },
        "visitor" : {
          "type" : "string",
          "example" : "AS Monaco"
        },
        "FT" : {
          "type" : "string",
          "example" : "0-1"
        },
        "hgoal" : {
          "type" : "integer",
          "format" : "int32",
          "example" : 0
        },
        "vgoal" : {
          "type" : "integer",
          "format" : "int32",
          "example" : 1
        },
        "division" : {
          "type" : "integer",
          "format" : "int32",
          "example" : 1
        },
        "tier" : {
          "type" : "integer",
          "format" : "int32",
          "example" : 1
        },
        "totgoal" : {
          "type" : "integer",
          "format" : "int32",
          "example" : 1
        },
        "goaldif" : {
          "type" : "integer",
          "format" : "int32",
          "example" : 1
        },
        "result" : {
          "type" : "string",
          "example" : "D",
          "enum" : [ "D", "H" ]
        }
      },
      "example" : {
        "date" : "1956-12-25T00:00:00.000Z",
        "totgoal" : 1,
        "vgoal" : 1,
        "FT" : "0-1",
        "home" : "Stade Rennes",
        "division" : 1,
        "result" : "D",
        "hgoal" : 0,
        "tier" : 1,
        "season" : 1956,
        "id" : 1,
        "goaldif" : 1,
        "visitor" : "AS Monaco"
      }
    },
    "CompetitionItem" : {
      "type" : "object",
      "required" : [ "competitionName", "countryName", "id", "seasonMax", "seasonMin" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32",
          "example" : 1
        },
        "countryName" : {
          "type" : "string",
          "example" : "France"
        },
        "competitionName" : {
          "type" : "string",
          "example" : "Ligue 1"
        },
        "seasonMin" : {
          "type" : "integer",
          "format" : "int32",
          "example" : 1933
        },
        "seasonMax" : {
          "type" : "integer",
          "format" : "int32",
          "example" : 2017
        }
      },
      "example" : {
        "seasonMin" : 1933,
        "competitionName" : "Ligue 1",
        "id" : 1,
        "countryName" : "France",
        "seasonMax" : 2017
      }
    }
  }
}